feat: Support Storybook 7 and docs pages
* Adapt the interface to be able to set up the custom sorting in Storybook 7 too, by a combination of changes in `main.js` and `preview.js`.
* Offer sorting the pages to groups by their type - docs and story.
* Add an article about how to migrate the configuration from the version 1 to 2.
* Migrate the project example.

BREAKING CHANGE: Only Storybook 7 is supported. If you use Storybook 6 or older, stay with the versions 1.x. Also, the minimum supported version of Node.js is 16. Except for the version requirements, the configuration of the custom sorting needs to be specified by a different interface. See the documentation about migration for more information.